Added POSTMAN collection, updated .gitignore, added README, LICENSE.

+++ src/main/java/com/agilebank/model/account/AccountDto.java
@@ -9,6 +9,6 @@ import lombok.NonNull;
 @AllArgsConstructor
 public class AccountDto {
   @NonNull private String id;
-  @NonNull private Long balance;
+  @NonNull private Double balance;
   @NonNull private Currency currency;
 }
