+++ src/main/java/com/agilebank/util/exceptions/InvalidAmountException.java
@@ -5,9 +5,9 @@ import lombok.Getter;
 @Getter
 public class InvalidAmountException extends RuntimeException {
 
-  private final Long amount;
+  private final Double amount;
 
-  public InvalidAmountException(Long amount) {
+  public InvalidAmountException(Double amount) {
     super("The amount of " + amount + " is invalid; please use a non-negative amount.");
     this.amount = amount;
   }
